% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{data_manip}
\alias{data_manip}
\title{Interactively manipulate master files.}
\usage{
data_manip(use_afs = TRUE, update = FALSE, data = NULL)
}
\arguments{
\item{use_afs}{Use master files from AFS}

\item{update}{Update AFS files before grabbing.}

\item{data}{Data to use (default NULL, and use AFS data)}
}
\description{
Mess around with data interactively (clean/reshape etc.)
}

