% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_genus_names.r
\name{extend_genus_names}
\alias{extend_genus_names}
\title{Expands abbreviated genus names}
\usage{
extend_genus_names(species_vector_by_group)
}
\arguments{
\item{species_vector_by_group}{a character vector containing complete
and abbreviated genus names.}
}
\value{
A character vector of species or genus names whom genus name has been expanded.
}
\description{
The function takes a strings of species names of the form \code{Bufo bufo, B. bufo},
and writes \code{Bufo bufo, Bufo bufo}.
}
\details{
The function only analyses texts, there is no call to a known species list.
When encontering an abbreviated genus name, the function replaces it with the last
known complete name beginning with the same letter.
}
\examples{
extend_genus_names(c("Bufo bufo", "B. bufo", "Buteo buteo", "B. buteo"))
[1] "Bufo bufo"   "Bufo bufo"   "Buteo buteo" "Buteo buteo"
}
\author{
Alban Sagouis
}
