% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_all}
\alias{get_all}
\title{Retrieve a vector of datasets from AFS}
\usage{
get_all(data, key)
}
\arguments{
\item{data}{dataset names}

\item{key}{key linking names to master file locations.}
}
\value{
list of data.tables names according to 'data'
}
\description{
Wrapper for sync.afs::get_data
}

