% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grep_vars}
\alias{grep_vars}
\title{Search datasets for variables matching patterns.}
\usage{
grep_vars(pattern, data, regex = TRUE, ...)
}
\arguments{
\item{pattern}{Vector of patterns to search for (can be regex).}

\item{data}{Vector of datasets from which to search for variable matches.}

\item{regex}{Logical whether to use regex or search for exact matches.}

\item{...}{Passed to \code{grep}}
}
\value{
List of datasets with matches.
}
\description{
Find variables matching patterns in different datasets.
}

