% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeniseNumberLength.r
\name{homogeniseNumberLength}
\alias{homogeniseNumberLength}
\title{Function transforming species numbers.}
\usage{
homogeniseNumberLength(species_vector_by_group)
}
\arguments{
\item{species_vector_by_group}{description}
}
\value{
A vector of the same length as species_vector_by_group.
}
\description{
Change the number of leading zeros to homogenise all numbers found in a vector of species
and morphospecies names.
The length of numbers will be homogenised among species names of a single study given in grouping_vector. If in a given study, the largest number has 3 digits, all numbers will get zeroes concatened before. 1 -> 001.
}
\examples{
Write an example.


}
\author{
Alban Sagouis
}
