% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nonEmpty}
\alias{nonEmpty}
\title{Remove null/empty elements from list}
\usage{
nonEmpty(lst, check_nested = TRUE)
}
\arguments{
\item{lst}{List to clean.}

\item{check_nested}{logical determines if nested lists should be unlisted.}
}
\value{
Cleaned list
}
\description{
remove nulls/empty values from list
}

