% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_names_correction.R
\name{site_names_correction}
\alias{site_names_correction}
\title{Homogenises site names
The function takes a strings of heterogeneous site names, corrects misprints and
homogenises spacing, capital letters and names.}
\usage{
site_names_correction(
  site_vector = NA,
  add_site_category_to_acronyms = FALSE,
  site_category = "Site",
  convert_to_ascii = FALSE
)
}
\arguments{
\item{site_vector}{a character vector containing site names to correct}

\item{add_site_category_to_acronyms}{Should a common name specified in \code{site_category}
should be added to sites only described by an acronym. Default to \code{FALSE}.}

\item{site_category}{Word added to accronyms if \code{add_site_category_to_acronyms = TRUE}.}

\item{convert_to_ascii}{Change the encoding of the vector to ASCII which replaces all
special characters to their closest ASCII equivalent:
accentuated characters such as Áêãçoàúü are converted to their basic equivalent Aeacoauu}
}
\value{
A character vector of same length as site_vector
}
\description{
Homogenises site names
The function takes a strings of heterogeneous site names, corrects misprints and
homogenises spacing, capital letters and names.
}
\details{
details
}
\examples{
extend_genus_names(c("Bufo bufo", "B. bufo", "Buteo buteo", "B. buteo"))
exmpl <- c('s2','1','site.3','forestPrime','forest,border','forest border','(border)')
data.frame(exmpl,
           corrected = site_names_correction(site_vector = exmpl,
                      add_site_category_to_acronyms = TRUE)
)
         exmpl     corrected
1            s2       Site s2
2             1        Site 1
3        site.3        Site.3
4   forestPrime  Forest Prime
5 forest,border Forest,border
6 forest border Forest border
7      (border)      (border)

}
\author{
Alban Sagouis
}
